requirements.txt
streamlit
PyMuPDF
Pillow
pytesseract
opencv-python
sentence-transformers
faiss-cpu
scikit-learn
numpy
google-generativeai
python-dotenv
asyncio
aiofiles